15 Hapus Tag Dan Modifikasi Route Middleware

+++ routes/web.php
@@ -37,13 +37,18 @@ Route::group(['middleware' => ['auth']], function () {
 
     });
 
-    Route::prefix('tags')->middleware('permission:create tags')->group(function () {
-        Route::get('create', [TagController::class, 'create'])->name('tags.create');
-        Route::post('create', [TagController::class, 'store']);
-        Route::get('table', [TagController::class, 'table'])->name('tags.table');
-        Route::get('{tag:slug}/edit', [TagController::class, 'edit'])->name('tags.edit');
-        Route::put('{tag:slug}/edit', [TagController::class, 'update']);
-        Route::delete('{tag:slug}', [TagController::class, 'destroy'])->name('tags.delete');
+    Route::prefix('tags')->group(function () {
+        Route::middleware('permission:create tags')->group(function() {
+            Route::get('table', [TagController::class, 'table'])->name('tags.table');
+            Route::get('create', [TagController::class, 'create'])->name('tags.create');
+            Route::post('create', [TagController::class, 'store']);
+        });
+
+        Route::middleware('permission:delete tags,edit tags')->group(function() {
+            Route::get('{tag:slug}/edit', [TagController::class, 'edit'])->name('tags.edit');
+            Route::put('{tag:slug}/edit', [TagController::class, 'update']);
+            Route::delete('{tag:slug}', [TagController::class, 'destroy'])->name('tags.delete')->middleware('permission:delete tag');
+        });
 
     });
 });
